+++ markets/bfp-market/contracts/modules/LiquidationModule.sol
@@ -29,6 +29,7 @@ contract LiquidationModule is ILiquidationModule {
     using Position for Position.Data;
 
     // --- Immutables --- //
+
     address immutable SYNTHETIX_CORE;
     address immutable SYNTHETIX_SUSD;
     address immutable ORACLE_MANAGER;
@@ -74,7 +75,7 @@ contract LiquidationModule is ILiquidationModule {
             uint256 liqKeeperFee
         )
     {
-        (int256 fundingRate, ) = market.recomputeFunding(oraclePrice);
+        (int128 fundingRate, ) = market.recomputeFunding(oraclePrice);
         emit FundingRecomputed(
             marketId,
             market.skew,
@@ -102,7 +103,7 @@ contract LiquidationModule is ILiquidationModule {
 
         emit MarketSizeUpdated(marketId, updatedMarketSize, updatedMarketSkew);
 
-        (uint256 utilizationRate, ) = market.recomputeUtilization(
+        (uint128 utilizationRate, ) = market.recomputeUtilization(
             oraclePrice,
             AddressRegistry.Data({
                 synthetix: ISynthetixSystem(SYNTHETIX_CORE),
